+++ server/src/index.ts
@@ -1,13 +1,12 @@
 import { config } from "config/config";
 
 import { configureDbSchemaValidation, connectToMongodb } from "./db/mongodb";
-import { modelDescriptors } from "./models/collections";
 import { server } from "./server";
 
 (async function () {
   try {
     await connectToMongodb(config.mongodb.uri);
-    await configureDbSchemaValidation(modelDescriptors);
+    await configureDbSchemaValidation();
 
     server.listen({ port: 5000, host: "0.0.0.0" }, function (err) {
       if (err) {
